Start fixing formula reference moves after shifing
